+++ CHANGELOG.md
@@ -8,6 +8,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 * Fix IE bug occuring when using step count > 1 and `max` on loop counter
 * Fix argument check for `seq` for programs with offset
 * Fix minimization/optimization loop
+* Fix program submission bug
 
 ## v25.6.27